move sponge funcs into separate package
